backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
openai==0.28.0
python-dotenv==1.0.0
gunicorn==20.1.0